First implementation of MC

+++ Python/Code/code_util.py
@@ -120,7 +120,7 @@ class DataManagement:
 
         if len(frames) == 2:
             print(type(frames[0]))
-            frames[0] = self.motion_compensation(np.array(frames[0]), np.array(frames[1]))
+            frames[0] = self.motion_compensation(np.array(frames[0]).astype(np.float32), np.array(frames[1]).astype(np.float32))
 
         for i in frames:
             cap.set(1, i)
@@ -155,8 +155,8 @@ class DataManagement:
         prev
     ):
 
-        curr_g = cv2.cvtColor(np.float32(curr), cv2.COLOR_YUV2GRAY_YUY2)
-        prev_g = cv2.cvtColor(np.float32(prev), cv2.COLOR_YUV2GRAY_YUY2)
+        curr_g = cv2.cvtColor(curr, cv2.COLOR_YUV2GRAY_YUY2)
+        prev_g = cv2.cvtColor(prev, cv2.COLOR_YUV2GRAY_YUY2)
 
         flow = cv2.calcOpticalFlowFarneback(prev_g, curr_g, None, 0.5, 3, 15, 3, 5, 1.2, 0)
